+++ DynamicsSimulator/src/Physics/Objects/MovableObject.h
@@ -10,19 +10,19 @@ private:
 	Vector2D m_Position, m_Speed, m_Acceleration;
 
 public:
-	double Mass;
+	float Mass;
 	Hitbox ObjectHitbox;
 	Graphics ObjectGraphics;
 
-	MovableObject(double X, double Y, double Mass, Hitbox ObjectHitbox);
-	MovableObject(double X, double Y, double Mass, Hitbox ObjectHitbox, Graphics ObjectGraphics);
-	MovableObject(double X, double Y, double Mass, Vector2D InitialSpeed, Hitbox ObjectHitbox, Graphics ObjectGraphics);
+	MovableObject(float X, float Y, float Mass, Hitbox ObjectHitbox);
+	MovableObject(float X, float Y, float Mass, Hitbox ObjectHitbox, Graphics ObjectGraphics);
+	MovableObject(float X, float Y, float Mass, Vector2D InitialSpeed, Hitbox ObjectHitbox, Graphics ObjectGraphics);
 #ifdef COPY_CONSTRUCTOR
 	MovableObject(const MovableObject& _mo);
 #endif
 
 	Vector2D GetPosition() const;
-	void SetPosition(double X, double Y);
+	void SetPosition(float X, float Y);
 	void SetPosition(Vector2D p);
 
 	Vector2D GetAcceleration() const;
